fix(proxy): #5905 — forward setPrototypeOf/deleteProperty through proxy targets with correct throw semantics
When a Proxy's target is itself a Proxy and the outer trap is missing/undefined/null,
the operation forwards down the chain. Several forwarding paths dropped the spec's
throw-on-false behavior or mis-reported non-configurable exotic properties as deletable:

- Object.setPrototypeOf(proxy, x) discarded the internal [[SetPrototypeOf]] boolean;
  a false result (non-extensible target reached through the chain) must throw a
  TypeError (Reflect.setPrototypeOf keeps returning the boolean). Fixes
  setPrototypeOf/trap-is-{missing,undefined}-target-is-proxy.

- Reflect.deleteProperty / delete never rejected an Array's non-configurable
  `length` or a plain function's non-configurable `prototype` — both are exotic
  own properties with no descriptor-table entry, so the delete silently reported
  success. Reject them in reflect_ordinary_delete_property_key and js_object_delete_field.

- delete proxy.key is lowered to the ProxyDelete HIR node, whose codegen returned
  js_proxy_delete's boolean directly and ignored strict mode; a strict-mode delete
  that resolves to false must throw. Route it through js_delete_result like the
  ordinary member-delete path. Fixes deleteProperty/trap-is-{missing,null,undefined}-target-is-proxy.

built-ins/Proxy 218/240 -> 223/240; no regressions in Reflect/Object/delete slices.
